feat: Move BreedsTableViewCell to UI core
